Move json and rdf nquad parsing under chunker.
